Fix division by 0 in TrapezoidalConstraint.calculateTranslation2d()
Clamping can reduce the desired velocity to the current velocity, which
isn't detected by the equality check at the beginning of the method. Add
a simple check that the clamped desired acceleration magnitude is
non-zero prior to division; if zero return the current velocity.

This fixes a regression introduced by
04fefe330bdd3b64360d0f5cfd09980e81b2aab2 (Fix TrapezoidalConstraint
bugs).