Change button bindings for chamber/intake
Use chamberNote() rather than startIntake() so that both the intake and
the chamber run.

Use cancelChambering() in addition to stopIntake() to stop both the
intake and the chamber. This is brittle; in the longer term we should
make it possible to reverse the intake without messing up the chamber
logic.