Tune note targeting
The Limelight data are remarkably noisy, so much of this change is
related to noise reduction. In particular, the median filter size is now
7 instead of 5, the angle tolerance is now 5 degrees instead of 3, and
some effort is expended to smooth the effects of spurious note detection
failures.

Fix a latent bug in note/speaker targeting angle velocity PID
initialization. Both PIDs were initialized with the I value for the D
parameter, but since both I and D were set to 0.0, this had no impact on
behavior.