Simplify Limelight subsystem
Remove distance calculation and area reporting, since 1) we didn't use
these values other than for dashboard reporting, and 2) the distance
calculation was incorrect. In order to accurately compute distance we
have to take the crosshairs position into account, which even were the
code corrected for this, would add brittleness/complication to
calibration.